SetParameters: use a larger buffer for the answer
The Protocol Data Structure is:
- 5 bytes long for a T=0 card
- 7 bytes long for a T=1 card

It may happen that the driver sends parameters for a T=0 card (5 bytes)
but the reader answers with a T=1 buffer (7 bytes).
The buffer is then too small for libusb and we get a LIBUSB_ERROR_OVERFLOW
error.

This should not happen in normal use but may happen with my HandlerTest
tool when I insert a T=1 card and ask for a T=0 communication. Some
reader (like the Cherry Smart Terminal 1150) may answer with a T=1
Protocol Data Structure.