feat: Update Ledger devices idProduct
Currently the app-openpgp from ledger tells users to patch the
Info.plist installed on their system for all devices to be supported.
Pushing an update on the readers description for this to not be
necessary anymore.

The idProduct values were extracted from
https://github.com/ledgerhq/ledger-secure-sdk
They are all written in lib_stusb/include/usbd_ledger.h

Signed-off-by: Francisco Freitas <francisco.freitas@ledger.fr>